+++ tf-infra/modules/ecs/variables.tf
@@ -58,3 +58,9 @@ variable "github_oauth_client_secret" {
   type      = string
   sensitive = true
 }
+
+variable "resend_api_key" {
+  type      = string
+  sensitive = true
+  default   = null
+}